python-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
svgwrite==1.4.3
requests==2.31.0
Werkzeug==3.0.1
